Catch exceptions to avoid crashes now.

+++ app/src/main/java/com/example/localcoin/CoinItem.kt
@@ -1,5 +1,5 @@
 package com.example.localcoin
 
-class CoinItem constructor(  public val name:String , public val qyt: Int){
+class CoinItem constructor(  public val name:String , public val qyt: Double){
 
 }
